+++ configs/self_forcing/kv_calib/wan_t2v_sf_sagequant_calib_14b.json
@@ -11,13 +11,18 @@
     "sample_shift": 5.0,
     "enable_cfg": false,
     "cpu_offload": false,
+    "dit_original_ckpt": "/data/nvme4/models/realtime-video/krea-realtime-video-14b.safetensors",
     "ar_config": {
         "local_attn_size": -1,
         "num_frame_per_chunk": 3,
         "timesteps_index": [0, 179, 358, 679],
+        "kv_quant": {
+            "calibrate": true,
+            "calib_path": "calib_kv.pt",
+            "quant_scheme": "sage",
+            "k_cache_type": "int8",
+            "v_cache_type": "fp8"
+        },
         "kv_offload": false
-    },
-    "dit_quantized": true,
-    "dit_quantized_ckpt": "/path/to/self_forcing_nvfp4.safetensors",
-    "dit_quant_scheme": "nvfp4"
+    }
 }
