+++ configs/self_forcing/wan_t2v_sf_kv_sagequant.json
@@ -4,20 +4,25 @@
     "text_len": 512,
     "target_height": 480,
     "target_width": 832,
-    "self_attn_1_type": "sage_attn2",
+    "self_attn_1_type": "sage_attn2_k_int8_v_fp8",
     "cross_attn_1_type": "sage_attn2",
     "cross_attn_2_type": "sage_attn2",
     "sample_guide_scale": 1,
     "sample_shift": 5.0,
     "enable_cfg": false,
     "cpu_offload": false,
+    "dit_original_ckpt": "/data/nvme4/gushiqiao/Self-Forcing/checkpoints/self_forcing_dmd.pt",
     "ar_config": {
         "local_attn_size": -1,
         "num_frame_per_chunk": 3,
         "timesteps_index": [0, 179, 358, 679],
+        "kv_quant": {
+            "calibrate": false,
+            "calib_path": "calib_kv.pt",
+            "quant_scheme": "sage",
+            "k_cache_type": "int8",
+            "v_cache_type": "fp8"
+        },
         "kv_offload": false
-    },
-    "dit_quantized": true,
-    "dit_quantized_ckpt": "/path/to/self_forcing_nvfp4.safetensors",
-    "dit_quant_scheme": "nvfp4"
+    }
 }
