+++ configs/self_forcing/wan_t2v_sf_kv_turboquant.json
@@ -11,13 +11,18 @@
     "sample_shift": 5.0,
     "enable_cfg": false,
     "cpu_offload": false,
+    "dit_original_ckpt": "/data/nvme4/gushiqiao/Self-Forcing/checkpoints/self_forcing_dmd.pt",
     "ar_config": {
         "local_attn_size": -1,
         "num_frame_per_chunk": 3,
         "timesteps_index": [0, 179, 358, 679],
+        "kv_quant": {
+            "calibrate": false,
+            "quant_scheme": "turboquant",
+            "codebook_dir": "/data/nvme4/gushiqiao/new/LightX2V/save_results/self_forcing/turboquant_codebooks",
+            "key_bits": 8,
+            "value_bits": 8
+        },
         "kv_offload": false
-    },
-    "dit_quantized": true,
-    "dit_quantized_ckpt": "/path/to/self_forcing_nvfp4.safetensors",
-    "dit_quant_scheme": "nvfp4"
+    }
 }
