+++ lightx2v/common/kvcache/quant.py
@@ -2,12 +2,6 @@ import os
 
 import torch
 import torch.distributed as dist
-
-try:
-    import fp4quant_cuda
-except ImportError:
-    fp4quant_cuda = None
-
 from loguru import logger
 
 from .kernel import *
@@ -331,145 +325,6 @@ class SageQuantRollingKVCachePool(RollingKVCachePool):
         v_fp8 = self._transpose_permute_v(v_fp8)
         v_scale = self._lookup_v_scale(layer_id).unsqueeze(0).contiguous()
         return v_fp8, v_scale
-
-
-class SageAttn3FP4RollingKVCachePool(RollingKVCachePool):
-    _BLK = 128
-    _V_GROUP = 16
-
-    def __init__(
-        self,
-        num_layers: int,
-        cache_size: int,
-        num_heads: int,
-        head_dim: int,
-        dtype: torch.dtype,
-        device: torch.device,
-        k_cache_type: str = "fp4",
-        v_cache_type: str = "fp4",
-        kv_offload: bool = False,
-    ) -> None:
-        assert k_cache_type in ["fp4"], f"Invalid k_cache_type: {k_cache_type}"
-        assert v_cache_type in ["fp4"], f"Invalid v_cache_type: {v_cache_type}"
-        if fp4quant_cuda is None:
-            raise ImportError("fp4quant_cuda is required for SageAttn3 FP4 KV cache.")
-        self._k_cache_type = k_cache_type
-        self._v_cache_type = v_cache_type
-        self._N_alloc = cdiv(int(cache_size), self._BLK) * self._BLK
-        super().__init__(
-            num_layers,
-            self._N_alloc,
-            num_heads,
-            head_dim,
-            dtype,
-            device,
-            kv_offload=kv_offload,
-        )
-
-    def _init_kv_buffer(self) -> None:
-        if self._kv_offload:
-            raise NotImplementedError("SageAttn3FP4RollingKVCachePool does not support kv_offload yet.")
-
-        L = self._num_layers
-        N = self._N_alloc
-        H = self._num_heads
-        D = self._head_dim
-        self._k_fp4 = torch.zeros(L, H, N, D // 2, dtype=torch.uint8, device=self._device)
-        self._k_scale = torch.zeros(L, H, N, D // 16, dtype=torch.float8_e4m3fn, device=self._device)
-        self._v_fp4 = torch.zeros(L, H, D, N // 2, dtype=torch.uint8, device=self._device)
-        self._v_scale = torch.zeros(L, H, D, N // 16, dtype=torch.float8_e4m3fn, device=self._device)
-        self._global_end = torch.zeros(L, dtype=torch.long, device=self._device)
-        self._local_end = torch.zeros(L, dtype=torch.long, device=self._device)
-
-    def store_kv(
-        self,
-        k: torch.Tensor,
-        v: torch.Tensor,
-        start_idx: int,
-        end_idx: int,
-        layer_id: int,
-    ) -> None:
-        chunk_len = int(end_idx - start_idx)
-        if chunk_len <= 0:
-            return
-        if k.size(0) != chunk_len or v.size(0) != chunk_len:
-            raise ValueError(f"SageAttn3FP4RollingKVCachePool.store_kv shape mismatch: chunk_len={chunk_len}, k={k.size(0)}, v={v.size(0)}.")
-        if start_idx % self._V_GROUP != 0 or chunk_len % self._V_GROUP != 0:
-            raise ValueError(f"SageAttn3 FP4 KV requires start/chunk aligned to {self._V_GROUP} (got start_idx={start_idx}, chunk_len={chunk_len}).")
-
-        padded_len = cdiv(chunk_len, self._BLK) * self._BLK
-        k_in = k.transpose(0, 1).unsqueeze(0).contiguous()  # [1, H, N, D]
-        v_in = v.transpose(0, 1).unsqueeze(0).contiguous()  # [1, H, N, D]
-
-        k_fp4 = torch.empty((1, self._num_heads, padded_len, self._head_dim // 2), device=self._device, dtype=torch.uint8)
-        k_scale = torch.empty((1, self._num_heads, padded_len, self._head_dim // 16), device=self._device, dtype=torch.float8_e4m3fn)
-        v_fp4 = torch.empty((1, self._num_heads, self._head_dim, padded_len // 2), device=self._device, dtype=torch.uint8)
-        v_scale = torch.empty((1, self._num_heads, self._head_dim, padded_len // 16), device=self._device, dtype=torch.float8_e4m3fn)
-
-        fp4quant_cuda.scaled_fp4_quant_permute(k_in, k_fp4, k_scale, 1)
-        fp4quant_cuda.scaled_fp4_quant_trans(v_in, v_fp4, v_scale, 1)
-
-        self._k_fp4[layer_id, :, start_idx:end_idx].copy_(k_fp4[0, :, :chunk_len])
-        self._k_scale[layer_id, :, start_idx:end_idx].copy_(k_scale[0, :, :chunk_len])
-        self._v_fp4[layer_id, :, :, start_idx // 2 : end_idx // 2].copy_(v_fp4[0, :, :, : chunk_len // 2])
-        self._v_scale[layer_id, :, :, start_idx // 16 : end_idx // 16].copy_(v_scale[0, :, :, : chunk_len // 16])
-
-    def k_cache(
-        self,
-        layer_id: int,
-        attn_start: int,
-        local_end: int,
-    ):
-        aligned_start = (attn_start // self._BLK) * self._BLK
-        kv_len = int(local_end - aligned_start)
-        padded_len = cdiv(kv_len, self._BLK) * self._BLK
-        k_fp4 = self._k_fp4[layer_id, :, aligned_start : aligned_start + padded_len].unsqueeze(0).contiguous()
-        k_scale = self._k_scale[layer_id, :, aligned_start : aligned_start + padded_len].unsqueeze(0).contiguous()
-        return k_fp4, k_scale, kv_len
-
-    def v_cache(
-        self,
-        layer_id: int,
-        attn_start: int,
-        local_end: int,
-    ):
-        aligned_start = (attn_start // self._BLK) * self._BLK
-        kv_len = int(local_end - aligned_start)
-        padded_len = cdiv(kv_len, self._BLK) * self._BLK
-        v_fp4 = self._v_fp4[layer_id, :, :, aligned_start // 2 : (aligned_start + padded_len) // 2].unsqueeze(0).contiguous()
-        v_scale = self._v_scale[layer_id, :, :, aligned_start // 16 : (aligned_start + padded_len) // 16].unsqueeze(0).contiguous()
-        return v_fp4, v_scale, kv_len
-
-    def roll_window(
-        self,
-        layer_id: int,
-        sink_tokens: int,
-        num_evicted: int,
-    ) -> None:
-        num_kept = int(self._local_end[layer_id].item()) - num_evicted - sink_tokens
-        if num_kept <= 0:
-            return
-        if sink_tokens % self._V_GROUP != 0 or num_evicted % self._V_GROUP != 0:
-            raise ValueError(f"SageAttn3 FP4 roll_window requires sink_tokens/num_evicted aligned to {self._V_GROUP} (got sink_tokens={sink_tokens}, num_evicted={num_evicted}).")
-
-        src_start = sink_tokens + num_evicted
-        src_end = src_start + num_kept
-        dst_start = sink_tokens
-        dst_end = dst_start + num_kept
-
-        self._k_fp4[layer_id, :, dst_start:dst_end].copy_(self._k_fp4[layer_id, :, src_start:src_end].clone())
-        self._k_scale[layer_id, :, dst_start:dst_end].copy_(self._k_scale[layer_id, :, src_start:src_end].clone())
-
-        self._v_fp4[layer_id, :, :, dst_start // 2 : dst_end // 2].copy_(self._v_fp4[layer_id, :, :, src_start // 2 : src_end // 2].clone())
-        self._v_scale[layer_id, :, :, dst_start // 16 : dst_end // 16].copy_(self._v_scale[layer_id, :, :, src_start // 16 : src_end // 16].clone())
-
-    def reset(self) -> None:
-        self._k_fp4.zero_()
-        self._k_scale.zero_()
-        self._v_fp4.zero_()
-        self._v_scale.zero_()
-        self._global_end.zero_()
-        self._local_end.zero_()
 
 
 class TurboQuantRollingKVCachePool(RollingKVCachePool):
